pyproject.toml
[tool.poetry]
name = "discord-hvz"
version = "0.3.0"
description = "A run-yourself Discord bot to manage Humans vs. Zombies games."
authors = ["ConnerHAnderson <conneranderson.dev@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
discord_hvz = "discord_hvz:main"
main = "discord_hvz:main"


[tool.poetry.dependencies]
python = ">=3.9.7,<3.11"
py-cord = "^2.2.0"
SQLAlchemy = "^1.4.32"
loguru = "^0.6.0"
"ruamel.yaml" = "^0.17.21"
google-api-python-client = "^2.42.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
regex = "^2022.3.15"
python-dotenv = "^0.20.0"
plotly = "^5.6.0"
pandas = "^1.4.1"
kaleido = "0.2.1"
pydantic-yaml = "^1.1"
tzdata = "*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
radon = "^5.1.0"
line-profiler = "^3.5.1"
pyinstaller = "^5.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
